(build) update npm script, use new logger in node script, rename file option to entry in node script

+++ build/src/cli/node/script.ts
@@ -1,13 +1,15 @@
 import path from 'node:path';
-import { logInfo, logError, runAsync } from '../helpers';
+import { logger, runAsync, pathExists } from '@ninjalib/util';
 import child_process from 'node:child_process';
 import { replaceTscAliasPaths } from 'tsc-alias';
 import fs from 'node:fs/promises';
 
+const log = logger('build:node');
+
 async function runtime(args) {
     const cwd = process.cwd();
 
-    const entryPoint = args.file ? path.join(cwd, args.file) : path.join(cwd, 'src/server/server.ts');
+    const entryPoint = args.entry ? path.join(cwd, args.entry) : path.join(cwd, 'src/server/server.ts');
 
     let subprocess: child_process.ChildProcess;
     if (args['--watch']) {
@@ -17,33 +19,30 @@ async function runtime(args) {
     }
 
     subprocess.on('error', (err) => {
-        logError(`failed to start subprocess: ${err}`);
+        log.error(`Failed to start subprocess: ${err}`);
     });
 
     subprocess.stdout?.on('data', (data) => {
-        logInfo(`stdout: ${data}`);
+        log.info(`Stdout: ${data}`);
     });
 
     subprocess.stderr?.on('data', (data) => {
-        logError(`stderr: ${data}`);
+        log.error(`Stderr: ${data}`);
     });
 
     subprocess.on('close', (code) => {
-        logInfo(`child process exited with code ${code}`);
+        log.info(`Child process exited with code ${code}`);
     });
 }
 
 async function build(args) {
     const cwd = process.cwd();
-    const entryPoint = args.file ? path.join(cwd, args.file) : path.join(cwd, 'src/server/server.ts');
+    const entryPoint = args.entry ? path.join(cwd, args.entry) : path.join(cwd, 'src/server/server.ts');
     const outDir = args.outDir ? path.join(cwd, args.outDir) : path.join(cwd, 'dist/node');
 
-    try {
-        // if outDir exists, remove old build
-        await fs.access(outDir, fs.constants.F_OK);
+    // if outDir exists, remove old build
+    if (await pathExists(outDir)) {
         await fs.rm(outDir, { recursive: true, force: true });
-    } catch (err) {
-        // outDir doesn't exist which is fine
     }
 
     const tempFilePath = path.join(cwd, 'node.tsconfig.json');
@@ -58,9 +57,10 @@ async function build(args) {
 
     try {
         try {
-            await runAsync(`tsc --project ${tempFilePath} && tsc-alias --project ${tempFilePath}`);
+            await runAsync(`tsc --project ${tempFilePath}`);
         } catch (err) {
-            // pass - most of the time this is just tsc complaining, the command doesn't actually "fail"
+            // pass - most of the time this is just tsc complaining,
+            // the command doesn't actually "fail"
         }
 
         // TODO?: there is a case this doesn't work, and doesn't replace the alias, if it meets these conditions:
@@ -71,9 +71,10 @@ async function build(args) {
         // - not very often is the project only going to have one folder to build from
         await replaceTscAliasPaths({ configFile: tempFilePath, resolveFullPaths: true });
     } catch (err) {
+        throw err;
+    } finally {
         await fs.unlink(tempFilePath);
     }
-    await fs.unlink(tempFilePath);
 }
 
 /**
@@ -88,7 +89,7 @@ export async function node(args: Record<string, string>) {
 }
 
 export const nodeOptions = {
-    '(optional) file': `The entry point for runtime/building (relative to cwd, defaults to CWD/src/server/server.ts)`,
+    '(optional) entry': `The entry point for runtime/building (relative to cwd, defaults to CWD/src/server/server.ts)`,
     '(optional) outDir': 'Where to build the project too (relative to cwd, defaults to CWD/dist/node)',
     '(optional) --watch': 'Enables auto reloading when files are changed',
     '(optional) --build': 'Builds the project instead of running node (cwd must have base tsconfig at root)',
